requirements.txt
amqp==5.1.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.9.1
async-timeout==4.0.2
attrs==23.2.0
billiard==3.6.4.0
boto3==1.35.43
botocore==1.35.43
bytecode==0.15.1
cachetools==5.5.2
cattrs==23.2.3
celery==5.2.7
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==2.0.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configcat-client==9.0.3
cron-descriptor==1.4.3
ddsketch==2.0.4
ddtrace==2.7.4
Deprecated==1.2.13
distro==1.9.0
Django==5.2.4
django-allow-cidr==0.7.1
django-celery-beat==2.6.0
django-cors-headers==3.7.0
django-timezone-field==6.1.0
djangorestframework==3.16.0
djangorestframework-jsonapi==4.2.1
enum-compat==0.0.3
envier==0.5.1
exceptiongroup==1.3.0
google-auth==2.38.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.2
importlib-metadata==6.11.0
inflection==0.5.1
iniconfig==2.1.0
ipaddress==1.0.23
jiter==0.5.0
jmespath==1.0.1
kombu==5.2.4
mssql-django==1.0rc1
openai==1.41.0
opentelemetry-api==1.23.0
packaging==25.0
pluggy==1.6.0
prompt-toolkit==3.0.30
protobuf==5.26.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
pyodbc==4.0.31
pyparsing==3.0.9
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.0
python-crontab==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.4
qualname==0.1.0
redis==5.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.3
semver==3.0.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==5.1.1
tenacity==9.0.0
tomli==2.2.1
tqdm==4.66.5
typing_extensions==4.14.1
tzdata==2024.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.18.1
